code updated

+++ modules/habits/router.js
@@ -20,10 +20,22 @@ router.post("/detail", async (req, res) => {
 	await habitsController.detail(req, res);
 });
 router.post("/update", async (req, res) => {
-	await habitsController.update(req, res);
+	if (req.role == "Administrator") {
+		await habitsController.update(req, res);
+	} else {
+		return res.status(400).send({
+			message: "You are not authorized to perform this action"
+		});
+	}
 });
 router.post("/delete", async (req, res) => {
-	await habitsController.delete(req, res);
+	if (req.role == "Administrator") {
+		await habitsController.delete(req, res);
+	} else {
+		return res.status(400).send({
+			message: "You are not authorized to perform this action"
+		});
+	}
 });
 router.post("/update/status", async (req, res) => {
 	if (req.role == "User") await habitsController.updateStatus(req, res);
